+++ backend/src/main/java/com/NBE_4_5_2/Team5/domain/user/service/email/service/DefaultTimeProvider.java
@@ -1,0 +1,12 @@
+package com.NBE_4_5_2.Team5.domain.user.service.email.service;
+
+import com.NBE_4_5_2.Team5.global.config.email.TimeProvider;
+import org.springframework.stereotype.Component;
+
+@Component
+public class DefaultTimeProvider implements TimeProvider {
+    @Override
+    public void sleep(long millis) throws InterruptedException {
+        Thread.sleep(millis);
+    }
+}
