+++ backend/src/main/java/com/NBE_4_5_2/Team5/global/config/email/TimeProvider.java
@@ -1,0 +1,5 @@
+package com.NBE_4_5_2.Team5.global.config.email;
+
+public interface TimeProvider {
+    void sleep(long millis) throws InterruptedException;
+}
